Fix a bug in PHP7's do_bind_inherited_class error printing code that causes segfaults
For the handler of declaring an inherited class.

That code attempted to treat the same zval for op2 as both an object and a
string on the same line to print an error message(It is actually a string)

Since the class entry was missing, the best thing that could be done
would probably be to print unknown instead of the object type.

The severity is minor, since Opcache is the only extension using this
function, and it doesn't look like this code path can be reached