[Proposal] Reduce the minimum size for packed arrays from 8 to 2
This can reduce the memory usage of some applications by ~16%
in applications that heavily use small arrays.

When the array size is unknown, continue to use the old array sizing choice to
avoid extra reallocations.
(seems better for benchmarks, though that may be sensitive to L1/L2 cache
and instruction cache)

Stop requiring that packed arrays have a size that is a power of 2.
There's no reason to for packed arrays,
since php isn't doing any hash lookups with bit operations.

https://nikic.github.io/2014/12/22/PHPs-new-hashtable-implementation.html
gives some background information on the `HashTable` implementation on
master.