[POC] Support calling global functions from constant expressions
- Currently, this treats all functions as global functions.
  Function resolution is straightforward to implement,
  I just haven't implemented it for a proof of concept.
- Argument unpacking isn't supported, but would be easy to implement.
- It turns out parameter defaults evaluate some constants every time,
  depending on whether the PHP value is immutable.

  It's likely we want to change this for expressions containing
  function calls, which can be more expensive than evaluating a regular constant AST.
  (e.g. reading files)
- This handles recursive definitions.
- Static calls with known classes (other than static::)
  may also be easy to implement.
- This also constrains function return values to be valid constants,
  (for what define() would accept)
  and throws an error otherwise.

  In the future, `const X = [my_call()->x];` may be possible.
- Variables aren't supported, because they depend on the declaration's scope,
  which may cease to exist.
- This throws an Error if a function call in a constant
  begins when it's still in progress.

Aside: https://github.com/TysonAndre/php-src/pull/10/files is an alternative
approach to a simple syntax to declaring constants with dynamic values,
which instead limits dynamic expressions of any type to `static const`.

-------

It turns out that function calls can already be invoked when evaluating a
constant, e.g. from php's error handler.
So it seems viable for PHP's to support this,
which this POC attempts to do.

(Imagine an error handler defining SOME_DYNAMIC_CONST123 to be a dynamic
value if a notice is emitted) for
`const X = [[]['undefined_index'], SOME_DYNAMIC_CONST123][1];`