Performance testing: Bind static variables as non-references.
The goal is to see if this is worth doing in opcache.
This creates a static variable that is a value instead of a reference,
subsequent assignments such as `$x = null;` do not modify the original
declaration.
(If $x was previously a reference before the `static` var statement,
it continues to be unset before the
expression on the right hand side is assigned to $x)

(e.g. `static $x : = [new ArrayObject(['foo' => 'bar'])]`

1. Locally benchmark **until implementing changes so that opcache can
   infer when it is safe to bind a `static $var` with $var as a
   non-reference** (e.g. only used by value, no foreach by ref, no $$ variable
   access or functions such as extract, no `require*/include*` statements,
   not in the top-level file scope, etc.).

   For example, `function (): int { static $var = 2; return $var; }`
   is likely much less efficient than a local variable
   but opcache doesn't attempt to optimize it.
2. Work out how to make the opcache changes correctly and safely